Move existing models to separate files
